A user can now delete their weight entry and also delete their profile which then in turn deletes all of their weight entries as well.

+++ app/controllers/application_controller.rb
@@ -19,23 +19,39 @@ class ApplicationController < ActionController::Base
     end
 
     def replace_weight_if_entered_today
-        new_weight = current_user.weights.last
-        previous_weight = current_user.weights[-2]
-        if new_weight.created_at.day == previous_weight.created_at.day
-            previous_weight.destroy
+        if current_user.weights[-2]
+            new_weight = current_user.weights.last
+            previous_weight = current_user.weights[-2]
+            if new_weight.created_at.day == previous_weight.created_at.day
+                previous_weight.destroy
+            end
         end
+        
     end
     
     def closer_to_goal
-        new_weight = current_user.weights.last
-        previous_weight = current_user.weights[-2]
-        new_difference_from_goal = (current_user.goal - new_weight.weight_input).abs()
-        previous_difference_from_goal = (current_user.goal - previous_weight.weight_input).abs()
-        if new_difference_from_goal < previous_difference_from_goal
-            "Great job! You are moving closer to your goal. Stay consistent and keep up the good work."
-        else
-            "It's okay, don't let this dishearten you. Keep focusing on your health and be patient your weight will be where you want it to be in time."
+        if current_user.weight
+            new_weight = current_user.weights.last 
+            previous_weight = current_user.weights[-2]
+            if new_weight
+                @new_difference_from_goal = (current_user.goal - new_weight.weight_input).abs()
+            else
+                @new_difference_from_goal = (current_user.goal - current_user.weight).abs()
+            end
+            
+            if previous_weight
+                @previous_difference_from_goal = (current_user.goal - previous_weight.weight_input).abs()    
+            else
+                @previous_difference_from_goal = (current_user.goal - current_user.weight).abs()
+            end
+            
+            if @new_difference_from_goal < @previous_difference_from_goal
+                "Great job! You are moving closer to your goal. Stay consistent and keep up the good work."
+            else
+                "It's okay, don't let this dishearten you. Keep focusing on your health and be patient your weight will be where you want it to be in time."
+            end
         end
+        
     end
 
 end
